refactor(graphql): rename graphql_defintions into graphql
