+++ lib/src/graphql.dart
